# Conflicts:
#	DataBase/InitialData/dbo.ApplicationSetting.Table.sql
#	DataBase/InitialData/dbo.Article.Table.sql
#	DataBase/InitialData/dbo.ArticleType.Table.sql
#	DataBase/InitialData/dbo.CMS_LayoutHtml.Table.sql
#	DataBase/InitialData/dbo.CMS_WidgetBase.Table.sql
#	DataBase/InitialData/dbo.CMS_Zone.Table.sql
#	DataBase/InitialData/dbo.DataArchived.Table.sql
#	DataBase/InitialData/dbo.DataDictionary.Table.sql
#	DataBase/InitialData/dbo.Forms.Table.sql
#	DataBase/InitialData/dbo.HtmlWidget.Table.sql
#	DataBase/InitialData/dbo.Language.Table.sql
#	DataBase/InitialData/dbo.Navigation.Table.sql
#	DataBase/InitialData/dbo.NavigationWidget.Table.sql
#	DataBase/InitialData/dbo.Permission.Table.sql
#	DataBase/InitialData/dbo.Product.Table.sql
#	DataBase/InitialData/dbo.ProductCategory.Table.sql
#	DataBase/InitialData/dbo.Roles.Table.sql
#	DataBase/InitialData/dbo.SectionContentCallToAction.Table.sql
#	DataBase/InitialData/dbo.Users.Table.sql
#	src/ZKEACMS.WebHost/compilerconfig.json

+++ src/ZKEACMS/Common/Service/CarouselWidgetService.cs
@@ -25,7 +25,9 @@ namespace ZKEACMS.Common.Service
         {
             _carouselItemService = carouselItemService;
         }
-        
+
+        public override DbSet<CarouselWidget> CurrentDbSet => DbContext.CarouselWidget;
+
         public override WidgetBase GetWidget(WidgetBase widget)
         {
             var carouselWidget = base.GetWidget(widget) as CarouselWidget;
@@ -49,6 +51,7 @@ namespace ZKEACMS.Common.Service
                         _carouselItemService.Add(new CarouselItemEntity
                         {
                             CarouselID = m.CarouselID,
+                            Title = m.Title,
                             CarouselWidgetID = item.ID,
                             TargetLink = m.TargetLink,
                             ImageUrl = m.ImageUrl,
@@ -76,7 +79,10 @@ namespace ZKEACMS.Common.Service
                     }
                     else if (m.ActionType == ActionType.Delete)
                     {
-                        _carouselItemService.Remove(m);
+                        if (m.ID > 0)
+                        {
+                            _carouselItemService.Remove(m);
+                        }                        
                     }
                     else
                     {
